refactor: useSearch
